add fix for id type with postgres

+++ src/main/java/net/lelyak/repository/MessageRepo.java
@@ -8,6 +8,6 @@ import java.util.List;
 /**
  * @author Nazar Lelyak.
  */
-public interface MessageRepo extends CrudRepository<Message, Long> {
+public interface MessageRepo extends CrudRepository<Message, Integer> {
     List<Message> findByTag(String tag);
 }
